Full set of results on the Pagie-1 problem without Tarpeian bloat control.
These use the koza2 function set as per the "standard" setup in ECJ,
which includes sine, cosine, exp, and ln. This differs from the function
set in the original Pagie paper and that used by the Hampshire group, both
of which just use +, -, *, and /. I'm still not sure what that means
for the problem and understanding our results, but I'm looking into it.