Added indentation to paragraphs.
The ACM format file doesn't seem to either indent the starts of paragraphs, or leave a blank line between paragraphs, which makes things pretty hard to read. I'd frankly prefer the blank line approach visually, but the indent uses up less space, and I think we're going to struggle to stay in 8 pages, so I went with indentation. I'm guessing that their style file is supposed to do _something_ with paragraphs, though, and we should probably try to get their way to work instead of hacking up our own.