Added a transformed version of the K-Landscape violin plot.
Taking the log of the y-scale didn't work because the fitnesses are between 0 and 1, so taking the log actually worsened the compression at the top. So I instead transformed the data by squaring it, which did seem to spread out the range in the way that David suggested. It's Figure 4 in this paper.