Added fitness vs. time plot for weak configuration with 10K pop.
I did a group of runs with the "weak" configuration (tournament size 7 and elitism of 0.1%) but with the large population size (10,240). Increasing the pop size definitely improves the performance by quite a lot (no surprise). The different bias levels are all essentially the same at the end of the 100 generations, but interestingly there is definitely a spread around 15-20 generations that is a really clean "more bias is better'' demonstration except for the fact that 0.75 and 1.0 are
essentially the same.