Removed Santa Fe Trail section and several Sine plots.
I removed the Santa Fe Trail section since we'd agreed that we were going to drop that for now. I also removed a bunch of plots in the Sine section that we weren't going to use (e.g., violin plots) and swapped the order of the Sine and Pagie-1 sections. I'm not entirely sure which order we're going to want those in, but the Pagie-1 problem is the more "interesting" problem for the community, so it seemed to make sense to start with it.