Added more text to the XO bias section, including the new root/non-root parent diagram.
I've added the new root/non-root parent diagram, and a few more paragraphs to the XO Bias section. I'm now thinking that a lot of what's in the start of that section really belongs in the introduction, leaving the XO bias section to focus on the particulars of how we implemented/used it.

I've also left some notes where we probably want some citations, but didn't actually interrupt myself to go find those. If someone's bored they could go hunting for reasonable things to put there.