Added questions/comments from today's video meeting.
I mostly added things as questions/comments in boldface rather than actually *doing* anything as we went through the paper today, so people should feel free to go through and act on things that are marked in the paper.

I also removed several packages that had accreted over time and that seems to have fixed both the indentation problem and the weird spacing issues in the References section.