Combined and moved the two "fitness over time" plots for the sine.
I combined those two "fitness over time" plots for the sine regression problem, and then moved the new plot up to Experimental Setup with the discussion of why we introduced the terms bias-effective and non-bias-effective.

I also moved the mention of our use of ECJ to the first paragraph of that section rather than having it dangle at the end of the section.