Set two figures to align to the top of pages.
After shrinking Figure 1, two of the later figures hopped onto a single "page" (column, really) on their own, which wasted a lot of space. I added `[t]` to each of them to push them to tops of pages and that cleaned that up a lot.

This also includes the current PDF.