Adding correct ACM style files.
Turns out that GECCO was linking to the wrong (out-of-date) ACM style files, and that the ACM (in their infinite wisdom) _didn't_ provide a notice at the top of the out-of-date page that we were looking at the wrong things, with a link to the correct page (https://www.acm.org/publications/article-templates/proceedings-template.html/).

This replaces the old, out-of-date style files with the new ones that actually work with the copyright notice.